Части кода разделены на модули.
