+++ js/data.js
@@ -1,0 +1,34 @@
+const PHOTOS_COUNT = 25;
+const PHOTOS_RANGE = [1, PHOTOS_COUNT];
+const LIKES_RANGE = [15, 200];
+const USERS_AVATAR_INDEXES = [1, 6];
+const COMMENTS_RANGE = [1, 20];
+
+const DESCRIPTIONS = [
+  'Закат',
+  'Лесной пейзаж',
+  'На берегу моря',
+  'В горах',
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const NAMES = [
+  'Иван',
+  'Мария',
+  'Сергей',
+  'Александр',
+  'Николай',
+  'Валентина',
+  'Андрей',
+  'Михаил',
+];
+
+export {PHOTOS_COUNT, PHOTOS_RANGE, LIKES_RANGE, USERS_AVATAR_INDEXES, COMMENTS_RANGE, DESCRIPTIONS, MESSAGES, NAMES};
