+++ js/util.js
@@ -1,0 +1,30 @@
+const randomNumberArray = (min, max) => {
+  const array = [];
+  for (let i = min; i <= max; i++) {
+    array.push(i);
+  }
+  return array;
+};
+
+const shuffleArray = (array) => {
+  for (let i = array.length - 1; i > 0; i--) {
+    const j = Math.floor(Math.random() * (i + 1));
+
+    const temp = array[i];
+    array[i] = array[j];
+    array[j] = temp;
+  }
+
+  return array;
+};
+
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+export {randomNumberArray, shuffleArray, getRandomInteger, getRandomArrayElement};
